removed compiler warning
git-svn-id: svn+ssh://svn.code.sf.net/p/gmat/code/trunk@1184 222415a0-0dc2-4184-858d-a27cde375526

+++ src/gui/mission/DecoratedTree.cpp
@@ -341,7 +341,7 @@ void DecoratedTree::DrawBoxes(wxTreeItemId id)
                     int boxDataNumber = i+lineNumber*boxCount;
 //                    MessageInterface::ShowMessage("Box data number %d\n", boxDataNumber);
                     
-                    if (boxDataNumber < boxData.size())
+                    if (boxDataNumber < (int)boxData.size())
                     {
                       if (boxData[boxDataNumber])
                       {
